# web/shared/openapi/admin-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "BergerHarbour.AdminApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/settings/business": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessSettingsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessSettingsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessSettingsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettingsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/seasons": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeasonDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeasonDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeasonDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/seasons/{id}": {
      "put": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/settings/blocked-periods": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockedPeriodDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockedPeriodDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockedPeriodDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedPeriodRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedPeriodRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedPeriodRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedPeriodDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedPeriodDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedPeriodDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/blocked-periods/{id}": {
      "put": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedPeriodRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedPeriodRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedPeriodRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedPeriodDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedPeriodDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedPeriodDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/settings/addons": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddonDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddonDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddonDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddonRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddonRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddonDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddonDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddonDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/addons/{id}": {
      "put": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddonRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddonRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddonDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddonDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddonDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/settings/email-templates/{key}": {
      "put": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EmailTemplateKey"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTemplateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTemplateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmailTemplateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailTemplateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailTemplateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/email-templates/{key}/test": {
      "post": {
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EmailTemplateKey"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddonDto": {
        "required": [
          "id",
          "version",
          "name",
          "description",
          "priceOnRequest",
          "prices",
          "quantityApplies",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "priceOnRequest": {
            "type": "boolean"
          },
          "prices": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AddonPricesDto"
              }
            ]
          },
          "quantityApplies": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "AddonPricesDto": {
        "required": [
          "midweek",
          "weekend",
          "week"
        ],
        "type": "object",
        "properties": {
          "midweek": {
            "type": "number",
            "format": "double"
          },
          "weekend": {
            "type": "number",
            "format": "double"
          },
          "week": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AddonRequest": {
        "required": [
          "version",
          "name",
          "description",
          "priceOnRequest",
          "prices",
          "quantityApplies",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "priceOnRequest": {
            "type": "boolean"
          },
          "prices": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AddonPricesDto"
              }
            ]
          },
          "quantityApplies": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "BlockedPeriodDto": {
        "required": [
          "id",
          "version",
          "name",
          "firstNight",
          "lastNight",
          "usesExtendedPaymentSchedule"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "firstNight": {
            "type": "string",
            "format": "date"
          },
          "lastNight": {
            "type": "string",
            "format": "date"
          },
          "usesExtendedPaymentSchedule": {
            "type": "boolean"
          }
        }
      },
      "BlockedPeriodRequest": {
        "required": [
          "version",
          "name",
          "firstNight",
          "lastNight",
          "usesExtendedPaymentSchedule"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "firstNight": {
            "type": "string",
            "format": "date"
          },
          "lastNight": {
            "type": "string",
            "format": "date"
          },
          "usesExtendedPaymentSchedule": {
            "type": "boolean"
          }
        }
      },
...
      "BusinessSettingsDto": {
        "required": [
          "version",
          "depositAmount",
          "minimumLeadTimeDays",
          "bookingsOpenUntil",
          "staffAlertEmail",
          "contactPhone",
          "contactEmail",
          "hireTermsUrl"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "depositAmount": {
            "type": "number",
            "format": "double"
          },
          "minimumLeadTimeDays": {
            "type": "integer",
            "format": "int32"
          },
          "bookingsOpenUntil": {
            "type": "string",
            "format": "date"
          },
          "staffAlertEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "hireTermsUrl": {
            "type": "string"
          }
        }
      },
...
      "EmailTemplateDto": {
        "required": [
          "key",
          "version",
          "subject",
          "htmlBody"
        ],
        "type": "object",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/EmailTemplateKey"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "subject": {
            "type": "string"
          },
          "htmlBody": {
            "type": "string"
          }
        }
      },
      "EmailTemplateKey": {
        "enum": [
          "BookingConfirmed",
          "StaffNewOnlineBooking",
          "PaymentDue",
          "PaymentReceived",
          "PreHireInstructions",
          "StaffBalanceOverdue",
          "StaffStandbySuperseded"
        ]
      },
      "EmailTemplateRequest": {
        "required": [
          "version",
          "subject",
          "htmlBody"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "subject": {
            "type": "string"
          },
          "htmlBody": {
            "type": "string"
          }
        }
      },
...
      "SeasonDto": {
        "required": [
          "id",
          "version",
          "name",
          "isDefault",
          "ranges"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "ranges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeasonRangeDto"
            }
          }
        }
      },
      "SeasonRangeDto": {
        "required": [
          "startDayMonth",
          "endDayMonth"
        ],
        "type": "object",
        "properties": {
          "startDayMonth": {
            "type": "string"
          },
          "endDayMonth": {
            "type": "string"
          }
        }
      },
      "SeasonRequest": {
        "required": [
          "version",
          "name",
          "ranges"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "ranges": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/SeasonRangeDto"
            }
          }
        }
      },
...
    }
  },
...
}